Extract common loadPhotos methods
