+++ app/src/main/java/me/ccrama/redditslide/Adapters/SubredditPosts.java
@@ -1,17 +1,11 @@
 package me.ccrama.redditslide.Adapters;
 
 import android.content.Context;
-import android.graphics.Bitmap;
 import android.os.AsyncTask;
 import android.util.Log;
-import android.view.View;
 import android.widget.Toast;
 
 import androidx.appcompat.app.ActionBar;
-import androidx.core.text.HtmlCompat;
-
-import com.nostra13.universalimageloader.core.assist.FailReason;
-import com.nostra13.universalimageloader.core.listener.ImageLoadingListener;
 
 import net.dean.jraw.http.NetworkException;
 import net.dean.jraw.models.Submission;
@@ -31,11 +25,11 @@ import me.ccrama.redditslide.Activities.SubredditView;
 import me.ccrama.redditslide.Authentication;
 import me.ccrama.redditslide.BuildConfig;
 import me.ccrama.redditslide.Constants;
-import me.ccrama.redditslide.ContentType;
 import me.ccrama.redditslide.Fragments.SubmissionsView;
 import me.ccrama.redditslide.HasSeen;
 import me.ccrama.redditslide.LastComments;
 import me.ccrama.redditslide.OfflineSubreddit;
+import me.ccrama.redditslide.PhotoLoader;
 import me.ccrama.redditslide.PostLoader;
 import me.ccrama.redditslide.PostMatch;
 import me.ccrama.redditslide.R;
@@ -89,173 +83,6 @@ public class SubredditPosts implements PostLoader {
             String subreddit) {
         this.subreddit = subreddit;
         loadMore(context, display, reset);
-    }
-
-    public void loadPhotos(List<Submission> submissions) {
-        for (Submission submission : submissions) {
-            String url;
-            ContentType.Type type = ContentType.getContentType(submission);
-            if (submission.getThumbnails() != null) {
-
-                if (type == ContentType.Type.IMAGE
-                        || type == ContentType.Type.SELF
-                        || (submission.getThumbnailType() == Submission.ThumbnailType.URL)) {
-                    if (type == ContentType.Type.IMAGE) {
-                        if (((!NetworkUtil.isConnectedWifi(c) && SettingValues.lowResMobile)
-                                || SettingValues.lowResAlways)
-                                && submission.getThumbnails() != null
-                                && submission.getThumbnails().getVariations() != null
-                                && submission.getThumbnails().getVariations().length > 0) {
-
-                            int length = submission.getThumbnails().getVariations().length;
-                            if (SettingValues.lqLow && length >= 3) {
-                                url = HtmlCompat.fromHtml(
-                                        submission.getThumbnails().getVariations()[2].getUrl(), HtmlCompat.FROM_HTML_MODE_LEGACY)
-                                        .toString(); //unescape url characters
-                            } else if (SettingValues.lqMid && length >= 4) {
-                                url = HtmlCompat.fromHtml(
-                                        submission.getThumbnails().getVariations()[3].getUrl(), HtmlCompat.FROM_HTML_MODE_LEGACY)
-                                        .toString(); //unescape url characters
-                            } else if (length >= 5) {
-                                url = HtmlCompat.fromHtml(submission.getThumbnails().getVariations()[
-                                        length
-                                                - 1].getUrl(), HtmlCompat.FROM_HTML_MODE_LEGACY).toString(); //unescape url characters
-                            } else {
-                                url = HtmlCompat.fromHtml(submission.getThumbnails().getSource().getUrl(), HtmlCompat.FROM_HTML_MODE_LEGACY)
-                                        .toString(); //unescape url characters
-                            }
-
-                        } else {
-                            if (submission.getDataNode().has("preview") && submission.getDataNode()
-                                    .get("preview")
-                                    .get("images")
-                                    .get(0)
-                                    .get("source")
-                                    .has("height")) { //Load the preview image which has probably already been cached in memory instead of the direct link
-                                url = submission.getDataNode()
-                                        .get("preview")
-                                        .get("images")
-                                        .get(0)
-                                        .get("source")
-                                        .get("url")
-                                        .asText();
-                            } else {
-                                url = submission.getUrl();
-                            }
-                        }
-
-
-                        ((Reddit) c.getApplicationContext()).getImageLoader()
-                                .loadImage(url, new ImageLoadingListener() {
-                                    @Override
-                                    public void onLoadingStarted(String imageUri, View view) {
-
-                                    }
-
-                                    @Override
-                                    public void onLoadingFailed(String imageUri, View view,
-                                            FailReason failReason) {
-
-                                    }
-
-                                    @Override
-                                    public void onLoadingComplete(String imageUri, View view,
-                                            Bitmap loadedImage) {
-
-                                    }
-
-                                    @Override
-                                    public void onLoadingCancelled(String imageUri, View view) {
-
-                                    }
-                                });
-
-                    } else if (submission.getThumbnails() != null) {
-
-                        if (((!NetworkUtil.isConnectedWifi(c) && SettingValues.lowResMobile)
-                                || SettingValues.lowResAlways)
-                                && submission.getThumbnails().getVariations().length != 0) {
-
-                            int length = submission.getThumbnails().getVariations().length;
-                            if (SettingValues.lqLow && length >= 3) {
-                                url = HtmlCompat.fromHtml(
-                                        submission.getThumbnails().getVariations()[2].getUrl(), HtmlCompat.FROM_HTML_MODE_LEGACY)
-                                        .toString(); //unescape url characters
-                            } else if (SettingValues.lqMid && length >= 4) {
-                                url = HtmlCompat.fromHtml(
-                                        submission.getThumbnails().getVariations()[3].getUrl(), HtmlCompat.FROM_HTML_MODE_LEGACY)
-                                        .toString(); //unescape url characters
-                            } else if (length >= 5) {
-                                url = HtmlCompat.fromHtml(submission.getThumbnails().getVariations()[
-                                        length
-                                                - 1].getUrl(), HtmlCompat.FROM_HTML_MODE_LEGACY).toString(); //unescape url characters
-                            } else {
-                                url = HtmlCompat.fromHtml(submission.getThumbnails().getSource().getUrl(), HtmlCompat.FROM_HTML_MODE_LEGACY)
-                                        .toString(); //unescape url characters
-                            }
-
-                        } else {
-                            url = HtmlCompat.fromHtml(submission.getThumbnails().getSource().getUrl(), HtmlCompat.FROM_HTML_MODE_LEGACY)
-                                    .toString(); //unescape url characters
-                        }
-
-                        ((Reddit) c.getApplicationContext()).getImageLoader()
-                                .loadImage(url, new ImageLoadingListener() {
-                                    @Override
-                                    public void onLoadingStarted(String imageUri, View view) {
-
-                                    }
-
-                                    @Override
-                                    public void onLoadingFailed(String imageUri, View view,
-                                            FailReason failReason) {
-
-                                    }
-
-                                    @Override
-                                    public void onLoadingComplete(String imageUri, View view,
-                                            Bitmap loadedImage) {
-
-                                    }
-
-                                    @Override
-                                    public void onLoadingCancelled(String imageUri, View view) {
-
-                                    }
-                                });
-
-                    } else if (submission.getThumbnail() != null && (submission.getThumbnailType()
-                            == Submission.ThumbnailType.URL
-                            || submission.getThumbnailType() == Submission.ThumbnailType.NSFW)) {
-
-                        ((Reddit) c.getApplicationContext()).getImageLoader()
-                                .loadImage(submission.getUrl(), new ImageLoadingListener() {
-                                    @Override
-                                    public void onLoadingStarted(String imageUri, View view) {
-
-                                    }
-
-                                    @Override
-                                    public void onLoadingFailed(String imageUri, View view,
-                                            FailReason failReason) {
-
-                                    }
-
-                                    @Override
-                                    public void onLoadingComplete(String imageUri, View view,
-                                            Bitmap loadedImage) {
-
-                                    }
-
-                                    @Override
-                                    public void onLoadingCancelled(String imageUri, View view) {
-
-                                    }
-                                });
-                    }
-                }
-            }
-        }
     }
 
     public ArrayList<String> all;
@@ -440,7 +267,7 @@ public class SubredditPosts implements PostLoader {
 
             if (!(SettingValues.noImages && ((!NetworkUtil.isConnectedWifi(c)
                     && SettingValues.lowResMobile) || SettingValues.lowResAlways))) {
-                loadPhotos(filteredSubmissions);
+                PhotoLoader.loadPhotos(c, filteredSubmissions);
             }
             if (SettingValues.storeHistory) {
                 HasSeen.setHasSeenSubmission(filteredSubmissions);
